Fix variabel type

+++ include/PenningTrap.hpp
@@ -6,5 +6,5 @@ class PenningTrap{
     private:
 
     public:
-        PenningTrap(double B0_in, double V0_in, int d_in, std::vector<Particle> particles_in);
+        PenningTrap(double B0_in, double V0_in, double d_in, std::vector<Particle> particles_in);
 };
